[X] Fix toList method for LazyList
There was an issue that in some use-cases elements
of the LazyList would not be evaluated in order
(for e.g. elem2 could be evaluated before elem1).

Turned out this was due to loop function inside
toList having rest as first argument which was
then repeatedly assigned with LazyList evaluated
tail. Head of LazyList was passed as part of the
second argument.
Function arguments are evaluated in order
left-to-right, therefore tail was evaluated before
head.
However, I am still not sure why this only started
happening when using map method. Maybe this is
somehow related to the fact that after applying map,
each LazyList element is a function returning another
function?