[5.8] Implement continually method in 2 variants for LazyList
The difference comes from strict vs non-strict parameter for
this method. If it is strict, then its value is evaluated
immediately but only once, therefore there no reason to
store the value in lazy val. If it is non-strict, then the
parameter's value would be evaluated for every element in
the LazyList. This is why in this case it is better to
memoize parameter's value and reference it recursively.